Add layer lock functionality
This locks a layer on, ignoring changes from momentary (MO) or one shot (OSL) layer changes.
This enables us to toggle on the current layer, without it being immediately turned off when the momentary key is released, or the one shot finishes.
For this key map, this means Func1 (CAPSLOCK) + Func2 (R Menu) toggles layer 1 on. Pressing any of these keys again will toggle it off.